StanWakefield/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
Pillow>=8.4.0
pytesseract>=0.3.8
tqdm>=4.62.0
matplotlib>=3.5.0 